+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - Added card-with-hyperlink-list pattern
 - Added horizontal-card pattern
 - Added icon-with-excerpt pattern
+- Added dswp-horizontal-card-large-img-left pattern
+- Added dswp-horizontal-card-large-img-right pattern
 
 ## 1.3.0 October 15, 2024